+++ packages/partitura/test/golden_test.dart
@@ -677,4 +677,17 @@ void main() {
       staffSpace: 9,
     );
   });
+
+  testWidgets('45 fingering numbers', (tester) async {
+    await golden(
+      tester,
+      '45_fingerings',
+      Score.simple(
+        timeSignature: TimeSignature.fourFour,
+        notes: 'c4:q=1 d4:q=2 e4:q=3 f4:q=4 | g4:q=5 e4:q=3 c4:q=1 r:q | '
+            'c4+e4+g4:h=1,3,5 r:h',
+      ),
+      staffSpace: 9,
+    );
+  });
 }